Move files to appropriate space
